Made the create dish method
