+++ src/main/java/Models/User.java
@@ -1,4 +1,4 @@
-package Model;
+package Models;
 
 import java.util.ArrayList;
 import java.util.List;
