+++ src/main/java/Persistens/ProductRepo.java
@@ -1,12 +1,10 @@
 package Persistens;
 
-import Model.Product;
-import org.w3c.dom.Text;
+import Models.Product;
 import util.TextUI;
 
 import java.sql.*;
 import java.util.ArrayList;
-import java.util.InputMismatchException;
 
 public class ProductRepo {
 
@@ -20,13 +18,13 @@ public class ProductRepo {
     public static ArrayList<Product> loadProducts() {
 
         try (Connection con = DriverManager.getConnection(connectionString)) {
-            TextUI.displayMsg("Connected to database");
 
             // The Statement is used to send SQL queries to the database. (SELECT, INSERT etc.)
             Statement stmt = con.createStatement();
             ResultSet rs = stmt.executeQuery(loadProduct);
 
             while (rs.next()) {
+                int Id = rs.getInt("Id");
                 String name = rs.getString("name");
                 String barcode = rs.getString("barcode");
                 double weight = rs.getFloat("weight");
@@ -36,7 +34,7 @@ public class ProductRepo {
                 int protein = rs.getInt("protein");
                 int fat = rs.getInt("fat");
 
-                Product product = new Product(name, barcode, weight, calories, carbs, sugar, protein, fat);
+                Product product = new Product(Id, name, barcode, weight, calories, carbs, sugar, protein, fat);
                 products.add(product);
             }
             stmt.close();
@@ -77,35 +75,7 @@ public class ProductRepo {
         } catch (SQLException e) {
             e.printStackTrace();
         }
-
-
         return false;
     }
 
-    public static void createProductByUser() {
-
-        boolean succes = false;
-
-        while (!succes)
-            try {
-                String name = TextUI.promptText("Enter the name of the product:");
-                String barcode = TextUI.promptText("Enter the barcode of the product:");
-                double weight = TextUI.promptDouble("Enter the weight of the product (in grams):");
-                int calorie = TextUI.promptInt("Enter the number of calories in the product:");
-                int carb = TextUI.promptInt("Enter the carbohydrate content of the product (in grams):");
-                int sugar = TextUI.promptInt("Enter the sugar content of the product (in grams):");
-                int protein = TextUI.promptInt("Enter the protein content of the product (in grams):");
-                int fat = TextUI.promptInt("Enter the fat content of the product (in grams):");
-
-                Product product = new Product(name, barcode, weight, calorie, carb, sugar, protein, fat);
-                saveProduct(product);
-                TextUI.displayMsg("Product added successfully");
-                succes = true;
-
-            } catch (InputMismatchException iME) {
-                TextUI.displayMsg("Invalid input! Please ensure you enter the correct type of value. Try again");
-                // Clear the invalid input from the buffer
-                TextUI.clearInputBuffer();
-            }
-    }
 }
