+++ src/main/java/Persistens/UserRepo.java
@@ -1,6 +1,6 @@
 package Persistens;
 
-import Model.User;
+import Models.User;
 import util.TextUI;
 
 import java.sql.*;
@@ -18,7 +18,6 @@ public class UserRepo {
     public static ArrayList<User> loadUsers(){
 
         try (Connection con = DriverManager.getConnection(connectionString)) {
-            TextUI.displayMsg("Connected to database");
 
             // The Statement is used to send SQL queries to the database. (SELECT, INSERT etc.)
             Statement stmt = con.createStatement();
@@ -45,7 +44,6 @@ public class UserRepo {
         String insertUserQuery = "INSERT INTO Users (userName, password) VALUES (?, ?)";
 
         try (Connection con = DriverManager.getConnection(connectionString)) {
-            TextUI.displayMsg("Connected to database");
 
             PreparedStatement pstmt = con.prepareStatement(insertUserQuery);
