+++ src/main/java/service/dishComparators/SortByCalories.java
@@ -1,6 +1,6 @@
 package service.dishComparators;
 
-import Model.Dish;
+import Models.Dish;
 import java.util.Comparator;
 
 public class SortByCalories implements Comparator<Dish> {
