+++ src/main/java/util/AppService.java
@@ -1,7 +1,7 @@
 package util;
 
-import Model.Product;
-import Model.User;
+import Models.Product;
+import Models.User;
 import Persistens.ProductRepo;
 
 import java.util.ArrayList;
